requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
streamlit>=1.23.1
python-dotenv>=1.0.0
groq>=0.4.0
nltk>=3.8.1
rank-bm25>=0.2.2
pydantic>=1.10.9
GitPython>=3.1.31
